[refactor] desktopApiMock をドメイン別ファイルへ分割 (WP-H7 PR2)
- mockRuntime.ts を新設: 共有可変状態(const レコード 14 + 再代入スカラー
  7 種)と状態依存ヘルパ 8 種を createMockRuntime() へ集約
- mocks/api/ に 8 ドメイン新設(posts / reactions / profileSocial /
  notifications / directMessages / liveGame / channels / connectivity)。
  各 createXxxMock(runtime) は Pick<DesktopApi, ...> を返し、
  const レコード・ヘルパは分割代入で bare 参照のまま、再代入スカラー
  (sequence / myProfile / notifications / recentReactions /
  discoveryConfig / communityNode*)のみ runtime.X で読み書き
- desktopApiMock.ts は合成のみ(1,559 → 33 行)。8 ドメインを
  単一 api へ spread するため this 経由の相互参照(channels の
  importChannelAccessToken → this.importFriendOnlyGrant 等)は不変
- oversized baseline から desktopApiMock を除去(10 → 9 エントリ)

挙動不変: Pick 返り型で全 81 メソッドの網羅と各シグネチャを型検査。
desktop-ui-check(mock ビルドの Storybook / ブラウザ E2E / visual)green。
