+++ xtask/oversized-baseline.json
@@ -5,10 +5,6 @@
       "path": "apps/desktop/src/styles/shell-phase1.css"
     },
     {
-      "lines": 1559,
-      "path": "apps/desktop/src/mocks/desktopApiMock.ts"
-    },
-    {
       "lines": 1155,
       "path": "apps/desktop/src/components/extended/MetaverseRoomPanel.tsx"
     },
